Add machinery to try and transform A ^ ((A ^ B) & M) into (A ^ ~M) | (B & M).
The former is MSVC's optimised idiom for bitfield assignment, the
latter is GCC's idiom.  The former causes Memcheck problems because it
doesn't understand that (in this complex case) XORing an undefined
value with itself produces a defined result.

Believed to be working but currently disabled.  To re-enable, change
if (0) to if (1) at line 6651.  Fixes, to some extent, and when
enabled, bug 344382.



git-svn-id: svn://svn.valgrind.org/vex@3097 8f6e269a-dfd6-0310-a8e1-e2731360e62c