Fix rounding mode check and instruction stxvl
In BE mode, the function FPU_rounding_mode_isOdd() has the assert
  vassert(mode->Iex.Const.con->Ico.U8 == 0x8);
The value was set using mkU32 but in BE mode the U8 maps to the upper
bits in the memory location not the lower bits.  The comparison was
fixed by changing the .U8 to .U32 to be consistent with how the field
was set.

The stxvl instruction called the 64-bit NOT not the 128-bit NOT when
calculating the store_val.

The stxvx instruction the temp word values were initialized I32 not I64.
Not sure why this wasn't caught on LE.

bugzilla  369175


git-svn-id: svn://svn.valgrind.org/vex@3252 8f6e269a-dfd6-0310-a8e1-e2731360e62c